# correctly unmarshal elements with a type tag