--- a/osm_test.go
+++ b/osm_test.go
@@ -183,7 +183,10 @@ func TestOSM_UnmarshalJSON(t *testing.T) {
 		  {"type":"way","id":456,"visible":false,"timestamp":"0001-01-01T00:00:00Z","nodes":[]},
 		  {"type":"relation","id":789,"visible":false,"timestamp":"0001-01-01T00:00:00Z","members":[]},
 		  {"type":"changeset","id":10,"created_at":"0001-01-01T00:00:00Z","closed_at":"0001-01-01T00:00:00Z","open":false},
-		  {"type":"user","id":16,"name":"","img":{"href":""},"changesets":{"count":0},"traces":{"count":0},"home":{"lat":0,"lon":0,"zoom":0},"languages":null,"blocks":{"received":{"count":0,"active":0}},"messages":{"received":{"count":0,"unread":0},"sent":{"count":0}},"created_at":"0001-01-01T00:00:00Z"},
+		  {"type":"user","id":16,"name":"","img":{"href":""},
+		   "changesets":{"count":0},"traces":{"count":0},"home":{"lat":0,"lon":0,"zoom":0},"languages":null,
+		   "blocks":{"received":{"count":0,"active":0}},"messages":{"received":{"count":0,"unread":0},"sent":{"count":0}},
+		   "created_at":"0001-01-01T00:00:00Z"},
 		  {"type":"note","id":15,"lat":0,"lon":0,"date_created":null,"date_closed":null,"comments":null}
 		]}`)
 
@@ -241,6 +244,25 @@ func TestOSM_UnmarshalJSON_Version(t *testing.T) {
 
 	if o.Version != "0.6" {
 		t.Errorf("incorrect version %v != 0.6", o.Version)
+	}
+}
+
+func TestOSM_UnmarshalJSON_Type(t *testing.T) {
+	data := []byte(`{
+		"version":0.6,"generator":"osm-go",
+		"elements":[
+			{"type":"relation","id":120,"timestamp":"0001-01-01T00:00:00Z","tags":{"type":"route"}},
+			{"tags":{"type":"route","other":"asdf"},"type":"relation","id":121,"timestamp":"0001-01-01T00:00:00Z"}
+		]}`)
+
+	o := &OSM{}
+	err := json.Unmarshal(data, &o)
+	if err != nil {
+		t.Fatalf("unmarshal error: %v", err)
+	}
+
+	if l := len(o.Relations); l != 2 {
+		t.Errorf("incorrect number of relations: %v", l)
 	}
 }