as described in https://github.com/paulmach/osm/issues/51 the "quick hack" to find the type does not work if the the element has `"tags":{"type":"route"}` like relations have.

This PR finds the tag using an unmarshal, it does impact performance. 
```
benchmark                            old ns/op     new ns/op     delta
BenchmarkChange_UnmarshalJSON-10     1099707       1331954       +21.12%

benchmark                            old allocs     new allocs     delta
BenchmarkChange_UnmarshalJSON-10     4880           5327           +9.16%

benchmark                            old bytes     new bytes     delta
BenchmarkChange_UnmarshalJSON-10     273049        252964        -7.36%
```

I was able to get it down to +7% speed but much more memory using `json.Decoder.Token`. I opted for the simpler approach as it allowed the plugging in of "customer unmarshallers". 

If performance is a concern consider using [jsoniter](github.com/json-iterator/go) or similar like
```
import (
  jsoniter "github.com/json-iterator/go"
  "github.com/paulmach/osm"
)

var c = jsoniter.Config{
  EscapeHTML:              true,
  SortMapKeys:             false,
  ValidateJsonRawMessage:  false,
  MarshalFloatWith6Digits: true,
}.Froze()

osm.CustomJSONMarshaler = c
osm.CustomJSONUnmarshaler = c
```